[d3d8] Properly initialize state block data
